svelte-migrate: updated files, update svelteki
