+++ package.json
@@ -25,7 +25,7 @@
     "@rollup/plugin-node-resolve": "^13.3.0",
     "@rollup/plugin-replace": "^4.0.0",
     "@sveltejs/adapter-node": "1.0.0-next.86",
-    "@sveltejs/kit": "1.0.0-next.405",
+    "@sveltejs/kit": "1.0.0-next.411",
     "eslint": "^8.22.0",
     "npm-run-all": "^4.1.5",
     "rollup": "^2.78.0",
