+++ src/routes/api/causal-diagram-widget-demo.html/+server.js
@@ -2,13 +2,14 @@ import { promises } from 'fs';
 import { baseUrl } from '$lib/passist';
 
 export async function GET() {
-	return {
-		headers:{
-			'content-type': 'text/html',
-		},
-		body:
-			(await promises.readFile(`static/api/_causal-diagram-widget-demo.html`))
+	return new Response(
+		(await promises.readFile(`static/api/_causal-diagram-widget-demo.html`))
 			.toString()
 			.replaceAll('https://passist.org', baseUrl),
-	};
+		{
+			headers: {
+				'content-type': 'text/html',
+			}
+		}
+	);
 }
