+++ src/routes/siteswap-alternatives/[siteswap]/+page.svelte
@@ -1,22 +1,11 @@
-<script context="module">
-	import { defaults } from '$lib/passist.mjs';
-
-	export async function load({ url, params }) {
-
-		return { props: {
-			siteswapInput: params.siteswap,
-			nJugglers: parseInt(url.searchParams.get('jugglers')) || defaults.nJugglers,
-			handsInput: url.searchParams.get('hands') || ''
-		} }
-	}
-</script>
-
 <script>
 	import SiteswapAlternativesPage from '$lib/SiteswapAlternativesPage.svelte';
 	import { useLocalStorage, siteswapAlternativesUrl } from '$lib/passist.mjs';
 	import { browser } from '$app/env';
 
-	export let siteswapInput, nJugglers, handsInput;
+	export let data;
+	$: ({siteswapInput, nJugglers, handsInput} = data);
+
 	let exclude = nJugglers == 2 ? '3 4.4' : '';
 
 	$:  useLocalStorage && siteswapInput && localStorage.setItem("siteswap-alternatives/siteswap", siteswapInput);
